Add files via upload
forgot some functions

+++ category.h
@@ -19,7 +19,7 @@ public:
     void addTransaction(QString category, QString name, QString date, double amount, QString type);
     //void editTransaction(); will be done later...ask about how edit will know what transaction to edit
     //void removeTransaction(); will be done later...ask about how remove will know what transaction to remove
-    double totalTransactions() const;           //adds up all of the amounts of every transaction in one category
+    int totalTransactions() const;           //adds up all of the amounts of every transaction in one category
 };
 
 #endif // SPENDINGS_H
